Ready! :D
Finally fixed error on K and order parameter not working: it was because the frequencies were called twice every time leading to non physical output. Now everything works as it should (so far!) with all the run embellishment and output previously done in the meanwhile. Check the README of the repository for the "to-do-s" list!